# feat: speedup import time using sys.platform