--- a/rich/color.py
+++ b/rich/color.py
@@ -1,5 +1,5 @@
-import platform
 import re
+import sys
 from colorsys import rgb_to_hls
 from enum import IntEnum
 from functools import lru_cache
@@ -15,7 +15,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from .text import Text
 
 
-WINDOWS = platform.system() == "Windows"
+WINDOWS = sys.platform == "win32"
 
 
 class ColorSystem(IntEnum):